fix: resolve CI issues
- Fix module conflict by moving tracing.rs to tracing/mod.rs
- Fix unused span.enter() error in grey-rpc
- Remove duplicate tracing.rs file
